refactor: expose component for external usage
