+++ src/custom/Carousel/index.tsx
@@ -1,0 +1,3 @@
+import Carousel from './Carousel';
+
+export { Carousel };
